pyproject.toml
[tool.poetry]
name = "ara"
version = "0.1.0"
description = "Autonomous Refactoring Agent - A LangGraph-based system for automated legacy code refactoring"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ara", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-google-genai = "^2.0.0"
google-generativeai = "^0.8.0"
libcst = "^1.1.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
psycopg2-binary = "^2.9.0"
asyncpg = "^0.29.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.0"
structlog = "^24.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pyright = "^1.1.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
